Create Image file
