+++ kadai1/asuforce/converter/converter.go
@@ -8,7 +8,6 @@ import (
 	"io"
 	"os"
 	"path/filepath"
-	"regexp"
 )
 
 // Converter struct
@@ -17,13 +16,6 @@ type Converter struct {
 	Files   []Image
 	FromExt string
 	ToExt   string
-}
-
-// Image information struct
-type Image struct {
-	path string
-	name string
-	ext  string
 }
 
 // Convert image functon
@@ -39,7 +31,7 @@ func (c *Converter) Convert(i Image) error {
 		return err
 	}
 
-	outputFile, err := os.Create(i.getFileName(c.ToExt))
+	outputFile, err := os.Create(i.GetFileName(c.ToExt))
 	if err != nil {
 		return err
 	}
@@ -57,7 +49,7 @@ func (c *Converter) Convert(i Image) error {
 func (c *Converter) CrawlFile(path string, info os.FileInfo, err error) error {
 	if checkExtension(filepath.Ext(path)) == ("." + c.FromExt) {
 		if !info.IsDir() {
-			c.Files = append(c.Files, newImage(path))
+			c.Files = append(c.Files, NewImage(path))
 		}
 	}
 	return nil
@@ -105,20 +97,4 @@ func (c *Converter) decodeImage(file io.Reader) (image.Image, error) {
 	}
 
 	return img, nil
-}
-
-func newImage(path string) Image {
-	ext := filepath.Ext(path)
-	rep := regexp.MustCompile(ext + "$")
-	name := filepath.Base(rep.ReplaceAllString(path, ""))
-
-	return Image{
-		path: path,
-		name: name,
-		ext:  ext,
-	}
-}
-
-func (i *Image) getFileName(ext string) string {
-	return i.name + "." + ext
 }
